Improved rage essence spawning conditions
Rage essence wasn't given enough time to exist in the world in order to be seen, and due to Doom's code for picking up items that overlapped, several of the drops would get absorbed despite having max values.

There's also now a brief delay on when the item can be picked up, giving a window for the player to see the drops before they move.